pyproject.toml
[tool.poetry]
name = "django-china-api"
version = "0.1.0"
description = ""
authors = ["jkc <1044500650@qq.com>"]

[tool.poetry.dependencies]
python = ">=3.7, <4"
django = "^3.2.8"
djangorestframework = "^3.12.4"
django-extensions = "^3.1.3"
drf-extensions = "^0.7.1"
django-model-utils = "^4.2.0"
django-imagekit = "^4.0.2"
pillow = "^8.4.0"
drf-spectacular = "^0.20.2"
djoser = "^2.1.0"
factory-boy = "^3.2.0"
django-environ = "^0.8.1"
mysqlclient = "^2.0.3"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
django-tree-queries = "^0.7.0"
django-tree-comments = "0.0.2"
django-notifications-plus = "^0.0.0"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.2"
pytest-django = "^4.4.0"
django-test-plus = "^2.2.0"
pytest = "^6.2.5"
isort = "^5.10.1"
black = "^21.11b1"
flake8 = "^4.0.1"
flake8-isort = "^4.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[tool.pytest.ini_options]
minversion = "6.0"
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = "tests.py test_*.py"
addopts = "--reuse-db"

[tool.isort]
profile = "black"
skip = ["migrations"]

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
  )/
)
'''
